AS-4: Adapta la lógica de negocios en UserService, inyecta UserService y UserRepository al construir el programa. Abstrae el modelo de registro a un ViewModel.
