+++ Locompro/Areas/Identity/Pages/Account/Register.cshtml
@@ -4,8 +4,6 @@
     ViewData["Title"] = "Register";
 }
 
-<h1>@ViewData["Title"]</h1>
-
 <div class="row">
     <div class="row justify-content-center">
         <div class="col-md-6 col-xl-4">
