+++ Locompro/Program.cs
@@ -46,12 +46,15 @@ void registerServices(WebApplicationBuilder builder)
     // Add DbContext using SQL Server
     builder.Services.AddDbContext<LocomproContext>(options =>
         options.UseSqlServer(builder.Configuration.GetConnectionString("LocomproContext") ?? throw new InvalidOperationException("Connection string 'LocomproContext' not found.")));
-    builder.Services.AddDefaultIdentity<User>(options => options.SignIn.RequireConfirmedAccount = true)
+    builder.Services.AddDefaultIdentity<User>(options => options.SignIn.RequireConfirmedAccount = false) /* add services to user here*/
         .AddEntityFrameworkStores<LocomproContext>();
+
     // Register repositories and services
     builder.Services.AddScoped<UnitOfWork>();
+    builder.Services.AddScoped<UserRepository>();
     builder.Services.AddScoped<StoreRepository>();
     builder.Services.AddScoped<StoreService>();
+    builder.Services.AddScoped<UserService>();
 }
